+++ spring-data-r2dbc/src/main/java/org/springframework/data/r2dbc/core/ReactiveDeleteOperation.java
@@ -113,7 +113,7 @@ public interface ReactiveDeleteOperation {
 		 * @return the number of affected rows; never {@literal null}.
 		 * @see Mono
 		 */
-		Mono<Integer> all();
+		Mono<Long> all();
 	}
 
 	/**
